Move differentiation module under lib/ and update imports
