+++ lib/differentiation.py
